refactor: API に関連する型定義を 1 ファイルに集約
