--- src/features/home/types/api.ts
@@ -1,7 +1,0 @@
-import { PopulationResponse } from "~/types/api";
-
-export type RawPopulationResponses = {
-  [n: number]: PopulationResponse["result"]["data"];
-};
-
-export type PopulationType = 0 | 1 | 2 | 3;
