+++ src/types/api.ts
@@ -20,3 +20,9 @@ export type PopulationResponse = {
     data: PopulationData[];
   };
 };
+
+export type RawPopulationResponses = {
+  [n: number]: PopulationResponse["result"]["data"];
+};
+
+export type PopulationType = 0 | 1 | 2 | 3;
